refactor: ProductRequest 생성
